Initial working WattWatchers fetch process
This is the start of the process for fetching data from the WattWatchers API.
At the moment it just reads a configuration file and then fetches the
list of WattWatcher devices available, and the data from each one.

The intent is to output to multiple sources - PVOutput.org and InfluxDB are the
ones I aim for.

Signed-off-by: Paul Wayper <paulway@mabula.net>