Retrieve usage data and upload to InfluxDB
This allows an 'influxdb' configuration section to define a connnection to an
InfluxDB server and database that will be used to store the usage data.  This
allows for a local storage of points across multiple days.

Eventually the InfluxDB server will become the main storage for data, and the
script will fetch data from WattWatchers into that and then upload from that
into PVOutput.

Signed-off-by: Paul Wayper <paulway@mabula.net>