Working complete fetch and storage of today's data
We now have a bit more of a coherent framework for fetching data from the
wattwatchers API and storing it either in a file or (untested) an InfluxDB
server.  This involves rearranging the data into channel -> timestamp ->
field, so that each channel's data can be fetched in one go rather than
distributing it across every timestamp.  This will allow us to check local
data timestamps in the future and only fetch missing times, as well as
output of a channel's worth of data to other systems.

Signed-off-by: Paul Wayper <paulway@mabula.net>