Lots of hacking mostly around getting data
There are kind of two competing projects here: one is a comprehensive system to
fetch data from WattWatchers and store it on the file system, in InfluxDB and
on PVOutput; the other is to grab the last complete five minutes' worth of data
from WattWatchers and upload that to PVOutput.

Obviously in the long term we aim to complete the first goal, and in doing so
we end up completing the second, the second is what's nagging at me.  And, to
be honest, trying to juggle all the complexities of fetching overlapping
samples of data, merge time ranges together, store and load them, know whether
they're complete or not, just got too difficult to do.  I needed a short term
goal.

So I got the latter working.  It does also have a few niceties: it will fill
in a power consumption reading as well if you set a 'consumption_channel'
setting in the pvoutput section of the `settings.yaml` file.  It will sleep
until about 30 seconds after the start of the five minute period, so the
WattWatcher devices have had a chance to communicate.  However, it's not
very tolerant of failure.

Signed-off-by: Paul Wayper <paulway@mabula.net>