Add fetch of today or a date and push to PVOutput
Gradually the pieces are coming together.

I got the 'push-pvoutput' mode working, and added the 'pvoutput-for-date'
mode, to fix the specific problem of having the machine that was running
the script in 'update-current' mode shut down.

'push-pvoutput' assumes today as the date range, whereas 'pvoutput-for-date'
requires a date via the `--date` or `-d` parameter (in 'YYYY-MM-DD' format).

Both modes then fetch any available WattWatcher data for the date range,
including possibly loading some data off the file system, and push this data
to PVOutput via its `addbatchstatus.jsp` page.

A `--verbose` (or `-v`) flag is also available to get more human information
about what the program is doing.

This is still a long way from a completely functional, modular system, but
the more I work on it the more coherent that picture starts to looks.

Signed-off-by: Paul Wayper <paulway@mabula.net>