First attempt at disk/web data merging
I've done some work on `get_usage_for_range` for when a day has already been
partly downloaded.  The fetch of the remainder of the day works, and I think
the metadata merge works, but the ranges fetched from disk seem to have
timestamps as strings rather than integers, which is confusing things.  Need
to work on this.

Signed-off-by: Paul Wayper <paulway@mabula.net>